Added ShapeRegion namespace with some shape functions.
