+++ WallsAndHoles/ellipsebrushtool.h
@@ -8,7 +8,7 @@ public:
     EllipseBrushTool(TileMapPreviewGraphicsItem *previewItem, QObject *parent = nullptr);
 
     /// Draws an ellipse.
-    QVector<QPoint> getShape(int dx, int dy) const override;
+    QRegion getShape(QPoint start, QPoint end) const override;
 };
 
 #endif // ELLIPSEBRUSHTOOL_H
